fix(astrology): resolve lint errors by moving types to shared file and correcting imports
